add Comment and secured for display of all projects

+++ View/project/project.html.php
@@ -1,13 +1,14 @@
 <?php
 
 use RedBeanPHP\R;
+
 //secured for user not connected
 if (!isset($_SESSION['user'])) {
     $_SESSION['error'] = "Veuillez vous connecter pour avoir accès à vos projets";
     (new App\Controller\AbstractController)->render('forms/login');
 }
-
-
+$userId = $_SESSION['user']['id'];
+$projectUserId = R::findAll('project', 'user_id=?', [$userId]);
 ?>
 
     <h2>Bienvenue sur votre espace de Time Tracking</h2>
@@ -16,40 +17,40 @@ if (!isset($_SESSION['user'])) {
     </div>
 
 <?php
-foreach ($data['project'] as $project) {
-    $task = R::find('task', 'project_id=?', [$project->id]);
-    ?>
-    <div class="content" id="<?= $project->id ?>">
-        <h2> <?= $project->project_title ?> </h2>
-        <div class="utils">
-            <p>
-                Création du projet <br> <br>
-                <i class="fa-regular fa-calendar-days"> <?= $project->project_date ?> </i>
-            </p>
-            <p>
-                Temps passé sur le projet : <br> <br>
-                <i class="fa-regular fa-clock"> <?= $project->projectTime ?> </i>
-            </p>
-        </div>
-        <div class="task">
-            <?php
-
-            foreach ($task as $tasks) { ?>
+foreach ($projectUserId as $project) {
+    if($project->user_id == $userId){
+        $task = R::find('task', 'project_id=?', [$project->id]);
+        ?>
+        <div class="content" id="<?= $project->id ?>">
+            <h2> <?= $project->project_title ?> </h2>
+            <div class="utils">
                 <p>
-                    <?= $tasks->taskName ?>
+                    Création du projet <br> <br>
+                    <i class="fa-regular fa-calendar-days"> <?= $project->project_date ?> </i>
                 </p>
+                <p>
+                    Temps passé sur le projet : <br> <br>
+                    <i class="fa-regular fa-clock"> <?= $project->projectTime ?> </i>
+                </p>
+            </div>
+            <div class="task">
+                <?php
+                foreach ($task as $tasks) { ?>
+                    <p>
+                        <?= $tasks->taskName ?>
+                    </p>
+                <?php } ?>
+            </div>
 
-            <?php } ?>
-        </div>
-
-        <div class="otherUtils">
+            <div class="otherUtils">
                 <span><a href="/index.php?c=project&a=delete-project&id=<?= $project->id ?>"><i
                                 class="fa-solid fa-trash suppress"></i></a></span>
-            <span><a href="/index.php?c=home&a=details-project&id=<?= $project->id ?>" title="Ajout de tâche / voir le détail">
+                <span><a href="/index.php?c=home&a=details-project&id=<?= $project->id ?>"
+                         title="Ajout de tâche / voir le détail">
                     <i class="fa-solid fa-eye"></i> Ajouter les tâches</a></span>
+            </div>
         </div>
-    </div>
-    <?php
+        <?php
+    }
 }
-
 ?>
